# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amo Market Portal API",
        "description": "This is the Portal API documentation for the Amo Market project.",
        "contact": {
            "email": "amomarket.info@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/portal/category": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories for shops near user",
                "description": "Returns product categories for shops within a 2 km radius of the user's current location.",
                "operationId": "getAllCategories",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude of the user's location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 12.9716
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude of the user's location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 77.5946
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Categories Retrieved Successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Fruits & Vegetables"
                                                    },
                                                    "shop_id": {
                                                        "type": "integer",
                                                        "example": 11
                                                    },
                                                    "content_image": {
                                                        "type": "string",
                                                        "example": "content_img.jpg"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "category_img.jpg"
                                                    },
                                                    "distance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.4
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing location data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Current location not provided."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No categories found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No categories found for shops within 3 km."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/portal/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get all service categories for nearby shops",
                "description": "Returns service-based categories for shops within a 2 km radius of the user's current location.",
                "operationId": "getAllServices",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude of the user's location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 12.9716
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude of the user's location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 77.5946
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Services retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Services Retrieved Successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Haircut"
                                                    },
                                                    "shop_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "content_image": {
                                                        "type": "string",
                                                        "example": "haircut_content.jpg"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "haircut.jpg"
                                                    },
                                                    "shop_type": {
                                                        "type": "string",
                                                        "example": "service"
                                                    },
                                                    "distance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.8
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing location data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Current location not provided."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No services found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No services found for shops within 3 km."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}